+++ .github/workflows/docs.yaml
@@ -7,7 +7,6 @@ on:
     paths:
       - "docs/**"
       - "docs-site/**"
-      - "tools/build-docs.mjs"
       - ".github/workflows/docs.yaml"
   workflow_dispatch:
