+++ package.json
@@ -19,12 +19,12 @@
   "license": "MIT",
   "author": "Brian (BMad) Madison",
   "scripts": {
-    "docs:build": "node tools/build-docs.mjs",
+    "docs:build": "node docs-site/scripts/build-docs.mjs",
     "docs:dev": "astro dev --root docs-site",
-    "docs:fix-links": "node tools/fix-doc-links.js",
+    "docs:fix-links": "node docs-site/scripts/fix-doc-links.js",
     "docs:preview": "astro preview --root docs-site",
-    "docs:validate-links": "node tools/validate-doc-links.js",
-    "docs:validate-sidebar": "node tools/validate-sidebar-order.js",
+    "docs:validate-links": "node docs-site/scripts/validate-doc-links.js",
+    "docs:validate-sidebar": "node docs-site/scripts/validate-sidebar-order.js",
     "format:check": "prettier --check \"**/*.{js,cjs,mjs,json,yaml}\"",
     "format:fix": "prettier --write \"**/*.{js,cjs,mjs,json,yaml}\"",
     "lint": "eslint . --max-warnings=0",
@@ -35,7 +35,7 @@
     "quality": "npm run format:check && npm run lint && npm run lint:py && npm run lint:md && npm run docs:build && npm test && npm run validate:refs && npm run validate:skills && npm run docs:validate-sidebar",
     "test": "npm run test:docs && npm run test:python",
     "test:docs": "node docs-site/test/test-site-url.mjs && node docs-site/test/test-rehype-plugins.mjs",
-    "test:implementation-model": "node test/test-validate-published-implementation-model.mjs",
+    "test:implementation-model": "node docs-site/test/test-validate-published-implementation-model.mjs",
     "test:python": "PYTHONDONTWRITEBYTECODE=1 uv run --frozen pytest",
     "validate:refs": "uv run tools/validate_file_refs.py --strict",
     "validate:skills": "uv run tools/validate_skills.py --strict"
